tests/test_canister/src/main.py
"""
Basilisk Toolkit Test Canister — minimal canister for integration testing.

Provides:
  - execute_code_shell: Execute Python code and return output (the core shell endpoint)
  - status: Health check

The frozen_stdlib_preamble automatically provides the in-memory filesystem (memfs).
ic-python-db provides the entity ORM for task/entity tests.
"""

import ic_python_db  # noqa: kept for module bundler dependency tracing
from basilisk import (
    Async,
    CallResult,
    GuardResult,
    Principal,
    StableBTreeMap,
    Tuple,
    ic,
    match,
    query,
    text,
    update,
)
from basilisk.canisters.management import (
    HttpResponse,
    HttpTransformArgs,
    management_canister,
)
from basilisk.db import Database

# ---------------------------------------------------------------------------
# Persistent database storage (survives canister upgrades)
# ---------------------------------------------------------------------------

storage = StableBTreeMap[str, str](memory_id=1, max_key_size=100, max_value_size=10000)
Database.init(db_storage=storage, audit_enabled=True)

# ---------------------------------------------------------------------------
# Persistent shell namespace (per principal)
# ---------------------------------------------------------------------------

_shell_ns_by_principal = {}


def guard_against_non_controllers() -> GuardResult:
    if ic.is_controller(ic.caller()):
        return {"Ok": None}
    return {
        "Err": "Not Authorized: only controllers of this canister may call this method"
    }


@update(guard=guard_against_non_controllers)
def execute_code_shell(code: str) -> str:
    """Execute Python code in a persistent namespace and return the output.

    Each caller principal gets its own isolated namespace that persists
    across calls. This is the core endpoint that basilisk shell uses.
    """
    import io
    import sys
    import traceback

    global _shell_ns_by_principal
    caller = str(ic.caller())
    if caller not in _shell_ns_by_principal:
        _shell_ns_by_principal[caller] = {"__builtins__": __builtins__}
        _shell_ns_by_principal[caller].update(
            {
                "ic": ic,
            }
        )
        _shell_ns_by_principal[caller]["basilisk"] = __import__("basilisk")
    ns = _shell_ns_by_principal[caller]

    stdout = io.StringIO()
    stderr = io.StringIO()
    sys.stdout = stdout
    sys.stderr = stderr

    try:
        exec(code, ns, ns)
    except Exception:
        traceback.print_exc()

    sys.stdout = sys.__stdout__
    sys.stderr = sys.__stderr__
    return stdout.getvalue() + stderr.getvalue()


@query
def status() -> str:
    """Health check endpoint."""
    return "ok"


@query
def whoami() -> str:
    """Return the caller's principal ID."""
    return str(ic.caller())


# ---------------------------------------------------------------------------
# HTTP outcall support
# ---------------------------------------------------------------------------


@query
def http_transform(args: HttpTransformArgs) -> HttpResponse:
    """Transform function for HTTP requests — removes headers for consensus."""
    response = args["response"]
    response["headers"] = []
    return response


@update(guard=guard_against_non_controllers)
def download_to_file(url: str, dest: str) -> Async[str]:
    """Download a file from a URL and save it to the canister filesystem.

    Makes an HTTP outcall via the IC management canister, then writes
    the response body (decoded as UTF-8) to *dest* on the in-memory
    filesystem.  Returns a human-readable status string.
    """
    http_result: CallResult[HttpResponse] = yield management_canister.http_request(
        {
            "url": url,
            "max_response_bytes": 2_000_000,  # IC limit
            "method": {"get": None},
            "headers": [
                {"name": "User-Agent", "value": "Basilisk/1.0"},
                {"name": "Accept-Encoding", "value": "identity"},
            ],
            "body": None,
            "transform": {
                "function": (ic.id(), "http_transform"),
                "context": bytes(),
            },
        }
    ).with_cycles(30_000_000_000)

    def _handle_ok(response: HttpResponse) -> str:
        try:
            content = response["body"].decode("utf-8")
        except UnicodeDecodeError as e:
            return f"Error: failed to decode response as UTF-8: {e}"
        import os

        parent = os.path.dirname(dest)
        if parent and parent != "/":
            os.makedirs(parent, exist_ok=True)
        with open(dest, "w") as f:
            f.write(content)
        return f"Downloaded {len(content)} bytes to {dest}"

    def _handle_err(err: str) -> str:
        return f"Download failed: {err}"

    return match(http_result, {"Ok": _handle_ok, "Err": _handle_err})
